genbarras/requirements.txt
defusedxml==0.7.1
fpdf2==2.5.6
Jinja2==3.0.3
MarkupSafe==2.1.5
pdfkit==1.0.0
Pillow==9.1.0
python-barcode==0.14.0
svg.path==6.3
